+++ src/miceFast_quantmodels.cpp
@@ -4,9 +4,13 @@
 #include <RcppArmadillo.h>
 #include "miceFast.h"
 
+#define UNUSED(expr) (void)(expr)
+
 //weighted linear regression
 
 arma::colvec fastLm_weighted(arma::colvec &y, arma::mat &X,arma::colvec &w, arma::mat &X1,int times) {
+
+  UNUSED(times);
 
   int n = X.n_rows, k = X.n_cols;
 
@@ -162,6 +166,8 @@ arma::colvec fastLm_noise(arma::colvec &y,arma::mat &X, arma::mat &X1,int times)
 
 arma::colvec fastLm_pred(arma::colvec &y, arma::mat &X, arma::mat &X1,int times) {
 
+  UNUSED(times);
+
   arma::colvec coef = arma::solve(X, y);//,arma::solve_opts::fast);//arma::inv(X.t()*X)*X.t()*y;
 
   return X1*coef;
@@ -172,13 +178,17 @@ arma::colvec fastLm_pred(arma::colvec &y, arma::mat &X, arma::mat &X1,int times)
 
 arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) {
 
+  UNUSED(times);
+
   arma::uvec vars = arma::find(arma::var(X)>0);
 
-  X = X.cols(vars);
+  arma::mat X_vol = X.cols(vars);
+
+  arma::mat X_means = arma::mean(X_vol,0);
 
   double tol = 1e-6;
 
-  int N = X.n_rows; int C = X.n_cols;
+  int N = X_vol.n_rows; int C = X_vol.n_cols;
 
   arma::vec un = arma::unique(y);
 
@@ -205,7 +215,7 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
 
   for(int i=0;i<N;i++) group_means_mat.row(i) = group_means.rows(arma::find(un == y(i)));
 
-  arma::mat Sw = arma::sqrt(arma::cov(X - group_means_mat));
+  arma::mat Sw = arma::sqrt(arma::cov(X_vol - group_means_mat));
 
   arma::vec scaling0 = 1/Sw.diag();
 
@@ -213,16 +223,19 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
 
   double fac =  1/(double)N;
 
-  arma::mat X0 = sqrt(fac) * (X - group_means_mat) * scaling;
+  arma::mat X0 = sqrt(fac) * (X_vol - group_means_mat) * scaling;
+
+  X_vol.clear();
 
   arma::mat input = X0.t()*X0;
+  X0.clear();
   arma::mat U;
   arma::mat V;
-  arma::vec s;
+  arma::vec s2;
 
-  arma::svd( U, s, V, input);
+  arma::svd( U, s2, V, input);
 
-  s = arma::sqrt(s);
+  arma::vec s = arma::sqrt(s2);
 
   arma::uvec proper = arma::find(s > tol);
 
@@ -242,11 +255,12 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
 
   fac = 1/(double) C;
 
-  X0 = arma::conv_to<arma::rowvec>::from(sqrt(((double) N * prior)*fac))* (group_means.each_row() - xb.t())  * scaling;
+  arma::mat X_s = arma::conv_to<arma::rowvec>::from(sqrt(((double) N * prior)*fac))* (group_means.each_row() - xb.t())  * scaling;
 
-  arma::svd( U, s, V, X0.t()*X0 );
+  arma::svd( U, s2, V, X_s.t()*X_s );
+  X_s.clear();
 
-  s = arma::sqrt(s);
+  s = arma::sqrt(s2);
 
   proper = arma::find(s > tol * s(0));
 
@@ -258,26 +272,26 @@ arma::colvec fastLda( arma::colvec &y,  arma::mat &X, arma::mat &X1, int times) 
 
   scaling = scaling * U.cols(proper);
 
-  arma::mat X_means = arma::mean(X,0);
-
   arma::vec dist_base = 0.5 * arma::mean(arma::pow(group_means * scaling,2),1) - arma::log(prior);
 
   N = X1.n_rows;
 
   //Prediction
 
-  X1 = X1.cols(vars);
+  arma::mat X1_vol = X1.cols(vars);
 
   arma::mat dist_base_mat =  arma::repmat(arma::trans(dist_base),N,1);
 
-  arma::mat dist_raw = dist_base_mat - X1 * scaling * arma::trans(group_means * scaling);
+  arma::mat dist_raw = dist_base_mat - X1_vol * scaling * arma::trans(group_means * scaling);
 
-  arma::mat dist_scale = arma::exp(-(dist_raw - arma::repmat(arma::min(dist_raw,1),1,group)));
+  dist_raw.each_col() -= arma::min(dist_raw,1);
+
+  dist_raw.transform( [](double val) { return (-std::exp(val)); } );
 
   arma::colvec pred(N);
 
   for(int i =0;i<N;i++){
-    int pos = arma::index_max(dist_scale.row(i));
+    int pos = arma::index_max(dist_raw.row(i));
     pred(i) = un(pos);
   }